requirements.txt
feedparser>=6.0,<7.0
jinja2>=3.1,<4.0
pyyaml>=6.0,<7.0
jsonschema>=4.20,<5.0
requests>=2.31,<3.0
